Addressing CVE-2021-22060 vulnerability.
Moderate severity
Vulnerable versions: >= 5.3.0, < 5.3.14
Patched version: 5.3.14
In Spring Framework versions 5.3.0 - 5.3.13, 5.2.0 - 5.2.18, and older unsupported versions, it is possible for a user to provide malicious input to cause the insertion of additional log entries. This is a follow-up to CVE-2021-22096 that protects against additional types of input and in more places of the Spring Framework codebase.